+++ junopoker/src/main/java/com/v1/junopoker/model/Player.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class Player {
     private String username;
-    private float chipCount;
+    private double chipCount;
     private Card[] holeCards = null;
     private Hand hand = null;
     private boolean inHand = false;
     private boolean isAllIn = false;
-    private float currentBet = 0;
+    private double currentBet = 0;
     private boolean isActive = false;
 }
